TPE-344 - refactor quark commands

See merge request ProtonMail/protonmail-ios!3157

+++ ProtonMail/ProtonMailUITests/Utils/Otp.swift
@@ -15,7 +15,7 @@ class Otp {
             let period = TimeInterval(30)
             let digits = 6
             let secret = Base32Decode(data: secret32)
-            var counter = UInt64(Date().timeIntervalSince1970 / period).bigEndian
+            var counter = UInt(Date().timeIntervalSince1970 / period).bigEndian
             let counterData = withUnsafeBytes(of: &counter) { Array($0) }
             let hash = HMAC<Insecure.SHA1>.authenticationCode(for: counterData, using: SymmetricKey(data: secret))
             var truncatedHash = hash.withUnsafeBytes { ptr -> UInt32 in
